+++ _examples/complete/main.tf
@@ -172,6 +172,7 @@ module "addons" {
   reloader                     = true
   external_dns                 = true
   redis                        = true
+  actions_runner_controller    = true
   prometheus                   = true
 
   # Jaeger deployment
@@ -210,6 +211,7 @@ module "addons" {
   reloader_helm_config                     = { values = [file("./config/reloader/override-reloader.yaml")] }
   external_dns_helm_config                 = { values = [file("./config/override-external-dns.yaml")] }
   redis_helm_config                        = { values = [file("./config/override-redis.yaml")] }
+  actions_runner_controller_helm_config    = { values = [file("./config/override-actions-runner-controller.yaml")] }  
   prometheus_helm_config                   = { values = [file("./config/override-prometheus.yaml")] }
 
   # -- Override Helm Release attributes
@@ -236,6 +238,7 @@ module "addons" {
   reloader_extra_configs                     = var.reloader_extra_configs
   external_dns_extra_configs                 = var.external_dns_extra_configs
   redis_extra_configs                        = var.redis_extra_configs
+  actions_runner_controller_extra_configs    = var.actions_runner_controller_extra_configs
   prometheus_extra_configs                   = var.prometheus_extra_configs
 
   # -- Custom IAM Policy Json for Addon's ServiceAccount